+++ src/main/java/com/example/comprehensivedegisn/adapter/repository/apart/ApartmentTransactionRepository.java
@@ -1,6 +1,6 @@
-package com.example.comprehensivedegisn.domain.repository;
+package com.example.comprehensivedegisn.adapter.repository.apart;
 
-import com.example.comprehensivedegisn.domain.ApartmentTransaction;
+import com.example.comprehensivedegisn.adapter.domain.ApartmentTransaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ApartmentTransactionRepository extends JpaRepository<ApartmentTransaction, Long> {
