+++ src/main/java/com/example/comprehensivedegisn/adapter/repository/interest/InterestRepository.java
@@ -1,6 +1,6 @@
-package com.example.comprehensivedegisn.domain.repository;
+package com.example.comprehensivedegisn.adapter.repository.interest;
 
-import com.example.comprehensivedegisn.domain.Interest;
+import com.example.comprehensivedegisn.adapter.domain.Interest;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
