+++ src/main/java/com/example/comprehensivedegisn/batch/api_client/KaKaoApiClientWithRoadName.java
@@ -5,8 +5,8 @@ import com.example.comprehensivedegisn.batch.kakao_map.KaKaoRestApiProperties;
 import com.example.comprehensivedegisn.batch.kakao_map.dto.ApartmentGeoRecord;
 import com.example.comprehensivedegisn.batch.kakao_map.dto.Documents;
 import com.example.comprehensivedegisn.batch.kakao_map.dto.LocationRecord;
-import com.example.comprehensivedegisn.domain.ApartmentTransaction;
-import com.example.comprehensivedegisn.domain.Gu;
+import com.example.comprehensivedegisn.adapter.domain.ApartmentTransaction;
+import com.example.comprehensivedegisn.adapter.domain.Gu;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpMethod;
