+++ src/main/java/com/example/comprehensivedegisn/batch/excel_write/dto/ExcelQueryRecord.java
@@ -1,9 +1,9 @@
 package com.example.comprehensivedegisn.batch.excel_write.dto;
 
-import com.example.comprehensivedegisn.domain.AddressUtils;
-import com.example.comprehensivedegisn.domain.ApartmentTransaction;
-import com.example.comprehensivedegisn.domain.DongEntity;
-import com.example.comprehensivedegisn.domain.Interest;
+import com.example.comprehensivedegisn.adapter.domain.AddressUtils;
+import com.example.comprehensivedegisn.adapter.domain.ApartmentTransaction;
+import com.example.comprehensivedegisn.adapter.domain.DongEntity;
+import com.example.comprehensivedegisn.adapter.domain.Interest;
 
 public record ExcelQueryRecord(
         // QuerydslNoOffsetIdPagingItemReader 작동하기 위해서는 ID 필드가 필요함
